+++ docs/global.js
@@ -1,0 +1,2 @@
+window.VERSION = "1.1.0";
+window.URL_PREFIX = location.pathname === "/" ? "" : location.pathname;