+++ docs/sidebar.en.js
@@ -1,0 +1,28 @@
+var I18N_EN = {
+    language: "English",
+    editLinkText: "Edit this page on GitHub",
+    nav: [
+        {
+            title: "Home",
+            link: "/en/README.md"
+        },
+        {
+            title: "Change Log",
+            link: "/CHANGELOG"
+        },
+        {
+            title: "GitHub",
+            link: "https://github.com/saekiraku/vscode-rainbow-fart"
+        }
+    ],
+    sidebar: [
+        {
+            title: "Summary",
+            link: "/en/README.md"
+        },
+        {
+            title: "Customize Voice Package",
+            link: "/en/voice-packages.md"
+        }
+    ]
+}