+++ docs/sidebar.zh.js
@@ -1,0 +1,27 @@
+var I18N_ZH = {
+    language: "简体中文",
+    editLinkText: "在 GitHub 上编辑此页",
+    nav: [{
+            title: "首页",
+            link: "/zh/README.md"
+        },
+        {
+            title: "更新日志",
+            link: "/CHANGELOG"
+        },
+        {
+            title: "GitHub",
+            link: "https://github.com/saekiraku/vscode-rainbow-fart"
+        }
+    ],
+    sidebar: [
+        {
+            title: "概述",
+            link: "/zh/README.md"
+        },
+        {
+            title: "定制语音包",
+            link: "/zh/voice-packages.md"
+        }
+    ]
+}