+++ CHANGELOG.md
@@ -4,6 +4,19 @@ All notable changes to the "vscode-rainbow-fart" extension will be documented in
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [1.1.0] - 2020-06-14
+
+### Added
+
+- Support import voice packages
+- New built-in voices for Chinese
+- I18N supports for English
+- Add document
+
+### Changed
+
+- New user interface
+
 ## [1.0.2] - 2020-06-01
 
 - Lower the required version of VSCode.